Separate traits into files
